Fixed horizon-sorting bug in A*. Allowed 8-directional movement in A*. Added grids for scout path distance

+++ src/Information/Geography/Pathfinding/Types/TilePath.scala
@@ -5,7 +5,7 @@ import Mathematics.Points.Tile
 case class TilePath(
   start     : Tile,
   end       : Tile,
-  distance  : Int,
+  distance  : Float,
   tiles     : Option[IndexedSeq[Tile]]) {
   
   def pathExists: Boolean = tiles.isDefined
